+++ crates/testing/src/spinning_task.rs
@@ -3,8 +3,8 @@ use std::{
     sync::Arc,
 };
 
-use async_lock::RwLock;
 use anyhow::Result;
+use async_lock::RwLock;
 use async_trait::async_trait;
 use futures::future::Either::{Left, Right};
 use hotshot::{traits::TestableNodeImplementation, types::EventType, HotShotInitializer};
@@ -23,9 +23,10 @@ use hotshot_types::{
 };
 use snafu::Snafu;
 
-use crate::test_task::{TestResult, TestTask, TestTaskState};
-
-use crate::test_task::{LateStartNode, Node, TestRunner};
+use crate::{
+    test_runner::{LateStartNode, Node, TestRunner},
+    test_task::{TestResult, TestTaskState},
+};
 
 /// convience type for state and block
 pub type StateAndBlock<S, B> = (Vec<S>, Vec<B>);
@@ -67,118 +68,118 @@ where
 {
     type Event = Event<TYPES>;
 
-    async fn handle_event(&mut self, (message, id): (Self::Event, usize)) -> Result<()> {
-         let Event { view_number, event } = message;
+    async fn handle_event(&mut self, (message, _id): (Self::Event, usize)) -> Result<()> {
+        let Event { view_number, event } = message;
 
-         if let EventType::Decide {
-             leaf_chain,
-             qc: _,
-             block_size: _,
-         } = event
-         {
-             let leaf = leaf_chain.first().unwrap().leaf.clone();
-             if leaf.get_view_number() > self.last_decided_leaf.get_view_number() {
-                 self.last_decided_leaf = leaf;
-             }
-         } else if let EventType::QuorumProposal {
-             proposal,
-             sender: _,
-         } = event
-         {
-             if proposal.data.justify_qc.get_view_number() > self.high_qc.get_view_number() {
-                 self.high_qc = proposal.data.justify_qc.clone();
-             }
-         }
+        if let EventType::Decide {
+            leaf_chain,
+            qc: _,
+            block_size: _,
+        } = event
+        {
+            let leaf = leaf_chain.first().unwrap().leaf.clone();
+            if leaf.get_view_number() > self.last_decided_leaf.get_view_number() {
+                self.last_decided_leaf = leaf;
+            }
+        } else if let EventType::QuorumProposal {
+            proposal,
+            sender: _,
+        } = event
+        {
+            if proposal.data.justify_qc.get_view_number() > self.high_qc.get_view_number() {
+                self.high_qc = proposal.data.justify_qc.clone();
+            }
+        }
 
-         // if we have not seen this view before
-         if self.latest_view.is_none() || view_number > self.latest_view.unwrap() {
-             // perform operations on the nodes
-             if let Some(operations) = self.changes.remove(&view_number) {
-                 for ChangeNode { idx, updown } in operations {
-                     match updown {
-                         UpDown::Up => {
-                             let node_id = idx.try_into().unwrap();
-                             if let Some(node) = self.late_start.remove(&node_id) {
-                                 tracing::error!("Node {} spinning up late", idx);
-                                 let node_id = idx.try_into().unwrap();
-                                 let context = match node.context {
-                                     Left(context) => context,
-                                     // Node not initialized. Initialize it
-                                     // based on the received leaf.
-                                     Right((storage, memberships, config)) => {
-                                         let initializer = HotShotInitializer::<TYPES>::from_reload(
-                                             self.last_decided_leaf.clone(),
-                                             TestInstanceState {},
-                                             None,
-                                             view_number,
-                                             self.high_qc.clone(),
-                                             Vec::new(),
-                                             BTreeMap::new(),
-                                         );
-                                         // We assign node's public key and stake value rather than read from config file since it's a test
-                                         let validator_config =
-                                             ValidatorConfig::generated_from_seed_indexed(
-                                                 [0u8; 32],
-                                                 node_id,
-                                                 1,
-                                                 // For tests, make the node DA based on its index
-                                                 node_id < config.da_staked_committee_size as u64,
-                                             );
-                                         TestRunner::add_node_with_config(
-                                             node_id,
-                                             node.networks.clone(),
-                                             memberships,
-                                             initializer,
-                                             config,
-                                             validator_config,
-                                             storage,
-                                         )
-                                         .await
-                                     }
-                                 };
+        // if we have not seen this view before
+        if self.latest_view.is_none() || view_number > self.latest_view.unwrap() {
+            // perform operations on the nodes
+            if let Some(operations) = self.changes.remove(&view_number) {
+                for ChangeNode { idx, updown } in operations {
+                    match updown {
+                        UpDown::Up => {
+                            let node_id = idx.try_into().unwrap();
+                            if let Some(node) = self.late_start.remove(&node_id) {
+                                tracing::error!("Node {} spinning up late", idx);
+                                let node_id = idx.try_into().unwrap();
+                                let context = match node.context {
+                                    Left(context) => context,
+                                    // Node not initialized. Initialize it
+                                    // based on the received leaf.
+                                    Right((storage, memberships, config)) => {
+                                        let initializer = HotShotInitializer::<TYPES>::from_reload(
+                                            self.last_decided_leaf.clone(),
+                                            TestInstanceState {},
+                                            None,
+                                            view_number,
+                                            self.high_qc.clone(),
+                                            Vec::new(),
+                                            BTreeMap::new(),
+                                        );
+                                        // We assign node's public key and stake value rather than read from config file since it's a test
+                                        let validator_config =
+                                            ValidatorConfig::generated_from_seed_indexed(
+                                                [0u8; 32],
+                                                node_id,
+                                                1,
+                                                // For tests, make the node DA based on its index
+                                                node_id < config.da_staked_committee_size as u64,
+                                            );
+                                        TestRunner::add_node_with_config(
+                                            node_id,
+                                            node.networks.clone(),
+                                            memberships,
+                                            initializer,
+                                            config,
+                                            validator_config,
+                                            storage,
+                                        )
+                                        .await
+                                    }
+                                };
 
-                                 let handle = context.run_tasks().await;
+                                let handle = context.run_tasks().await;
 
-                                 // Create the node and add it to the state, so we can shut them
-                                 // down properly later to avoid the overflow error in the overall
-                                 // safety task.
-                                 let node = Node {
-                                     node_id,
-                                     networks: node.networks,
-                                     handle,
-                                 };
-                                 node.handle.hotshot.start_consensus().await;
+                                // Create the node and add it to the state, so we can shut them
+                                // down properly later to avoid the overflow error in the overall
+                                // safety task.
+                                let node = Node {
+                                    node_id,
+                                    networks: node.networks,
+                                    handle,
+                                };
+                                node.handle.hotshot.start_consensus().await;
 
-                                 self.handles.write().await.push(node);
-                             }
-                         }
-                         UpDown::Down => {
-                             if let Some(node) = self.handles.write().await.get_mut(idx) {
-                                 tracing::error!("Node {} shutting down", idx);
-                                 node.handle.shut_down().await;
-                             }
-                         }
-                         UpDown::NetworkUp => {
-                             if let Some(handle) = self.handles.read().await.get(idx) {
-                                 tracing::error!("Node {} networks resuming", idx);
-                                 handle.networks.0.resume();
-                                 handle.networks.1.resume();
-                             }
-                         }
-                         UpDown::NetworkDown => {
-                             if let Some(handle) = self.handles.read().await.get(idx) {
-                                 tracing::error!("Node {} networks pausing", idx);
-                                 handle.networks.0.pause();
-                                 handle.networks.1.pause();
-                             }
-                         }
-                     }
-                 }
-             }
+                                self.handles.write().await.push(node);
+                            }
+                        }
+                        UpDown::Down => {
+                            if let Some(node) = self.handles.write().await.get_mut(idx) {
+                                tracing::error!("Node {} shutting down", idx);
+                                node.handle.shut_down().await;
+                            }
+                        }
+                        UpDown::NetworkUp => {
+                            if let Some(handle) = self.handles.read().await.get(idx) {
+                                tracing::error!("Node {} networks resuming", idx);
+                                handle.networks.0.resume();
+                                handle.networks.1.resume();
+                            }
+                        }
+                        UpDown::NetworkDown => {
+                            if let Some(handle) = self.handles.read().await.get(idx) {
+                                tracing::error!("Node {} networks pausing", idx);
+                                handle.networks.0.pause();
+                                handle.networks.1.pause();
+                            }
+                        }
+                    }
+                }
+            }
 
-             // update our latest view
-             self.latest_view = Some(view_number);
-         }
+            // update our latest view
+            self.latest_view = Some(view_number);
+        }
 
         Ok(())
     }
